### 2019-10-13 #1805 @whedon
[PRE REVIEW]: fathon: A Python package for a fast computation of detrendend fluctuation analysis and related algorithms
**Submitting author:** @stfbnc (<a href="http://orcid.org/0000-0001-6729-8655">Stefano Bianchi</a>)
**Repository:** <a href="https://github.com/stfbnc/fathon.git" target ="_blank">https://github.com/stfbnc/fathon.git</a>
**Version:** v0.1
