### 2019-10-13 #1805 @labarba
Check if removing the extra `---` around the metadata will make your paper compile. After the fix in the Markdown file, use `@whedon generate pdf` alone in a comment here to recompile.
