### 2019-10-15 #1805 @mbobra
@stfbnc Do you have any suggestions for potential reviewers from [this list of JOSS referees](https://bit.ly/joss-reviewers)?
