### 2019-10-15 #1805 @ali-ramadhan
Thanks for the email @mbobra. More than happy to review this submission.

I've encountered DFA in a few papers I've read (mostly applied to climate science) but I'm no expert, so hoping to learn more by reviewing fathon.
