### 2019-11-06 stfbnc/fathon#1 @ali-ramadhan
Difficulty installing fathon on Linux
I installed `gsl` through `pacman` but ran into a few issues.

The command given in step 3 of the README seems incorrect.
```
$ python setup.py build_ext -- =/usr/include/gsl/ -- =/usr/lib/ --inplace
zsh: /usr/include/gsl/ not found
```
so I changed it to
```
$ python setup.py build_ext -- /usr/include/gsl/* -- /usr/lib/ --inplace
Traceback (most recent call last):
  File "setup.py", line 2, in <module>
    from Cython.Build import cythonize
ModuleNotFoundError: No module named 'Cython'
```
so apparently I had to install install `Cython`. Googling brought me to https://github.com/CellProfiler/centrosome/issues/78 so it was simple to fix and now I think I got it to start compiling but I ran into another issue
```
$ python setup.py build_ext -- /usr/include/gsl/* -- /usr/lib/ --inplace
Compiling src/dfa.pyx because it changed.
Compiling src/dcca.pyx because it changed.
Compiling src/mfdfa.pyx because it changed.
Compiling src/ht.pyx because it changed.
[1/4] Cythonizing src/dcca.pyx
[2/4] Cythonizing src/dfa.pyx
[3/4] Cythonizing src/ht.pyx
[4/4] Cythonizing src/mfdfa.pyx
invalid command name '/usr/include/gsl/gsl_blas.h'
```

Apologies if I'm just doing something stupid. Just hoping to install fathon and try playing around with it.

Thanks so much!
