### 2019-11-06 stfbnc/fathon#2 @ali-ramadhan
Possible to have automated installation through pip or conda?
I wonder if providing a Python package installable via `pip` or `conda` would make fathon more accessible and easier to install. I'm not super familiar with either but because of the C depdencies, presumably `conda` is the better option?

GSL seems to be a pre-requisite which is available through most Linux package managers. Installing fathon on Windows might prove difficult and it looks like you might need to install all of Cygwin just to get GSL to install fathon. I might mention this as most Windows users have no idea what Cygwin is.

I was not able to install fathon myself.

I believe the installation instructions are unfriendly. fathon seems much harder to install for me than the vast majority of Python packages, many of them which depend on C libraries too.

Besides having to manually install or compile GSL, you also need to know where include and library files are placed on your operating system, and you need to know the "standard location of Python packages" for your particular OS and Python environment combination.

I think this is a huge barrier to usability and most Python users won't be able to install fathon.
