### 2019-11-06 #1828 @mbobra
> My major comments are:
>
> 1. fathon is difficult to install. I was actually not able to install it. It could greatly benefit from an automated package management solution.
>
>2. Currently documentation is provided via a single pdf file. I think generating documentation using a package like Sphinx would provide potential users with searchable documentation that gets updated alongside the code and integrates in the great docstrings that are already in the source code.
>
>3. Right now the burden is on the user to run manual tests to verify that fathon works, but I think this approach cannot scale as improvements are made and other people decide to contribute. I think an automated testing solution would greatly improve the situation here.
>
>
> Maybe @mbobra can confirm whether these three points are necessary for publication in JOSS. I do think they would improve fathon and make the package much more accessible to potential users.

Thanks for your detailed review, @ali-ramadhan. Here are my replies:

1. According to the JOSS review guidelines, [instructions for installing the software should be clear](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). An automated package management solution is ideal, but not required. However, the authors must provide clear installation instructions that work as well as a clear list of dependencies.

2. According to the JOSS review guidelines, the [documentation must to include four items](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation): a statement of need, installation instructions, example usage, and API documentation. The first three are pretty clear; the fourth is a bit subjective (and up to the reviewer to decide), but the core API must be documented in a reasonable way.

3. According to JOSS' review guidelines, [the reviewer must be able to check to see if the software actually works](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). An automated test suite is ideal, but documented manual steps are okay too. But there must be some way for the reviewer to objectively assess whether the software works.

Does this help, @ali-ramadhan?
