### 2019-11-07 stfbnc/fathon#2 @stfbnc
Ok, I'll try to make fathon available via `pip`, or at least include GSL directly in the package.

I've realized now that is not clear from the README, but for now the package is intended for macOS and Linux only. I do not have access to a Windows machine.
