### 2019-11-07 #1828 @ali-ramadhan
Thank you @mbobra, that's helpful! Sorry it's my first review but I should have familiarized myself with the review guidelines more closely.

> 1. According to the JOSS review guidelines, [instructions for installing the software should be clear](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). An automated package management solution is ideal, but not required. However, the authors must provide clear installation instructions that work as well as a clear list of dependencies.

I would be happy with clear installation instructions that can be followed by people without having to spend too much time figuring out where GSL include/library files are located and where Python packages are usually installed.

> 2. According to the JOSS review guidelines, the [documentation must to include four items](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation): a statement of need, installation instructions, example usage, and API documentation. The first three are pretty clear; the fourth is a bit subjective (and up to the reviewer to decide), but the core API must be documented in a reasonable way.

I think fathon satisfies the first three and does technically satisfy the fourth as functions in the source code are documented with good docstrings. I feel like it doesn't count as API documentation though unless the docstrings are compiled into documentation, but maybe having docstrings in the source code meets the JOSS requirements.

> 3. According to JOSS' review guidelines, [the reviewer must be able to check to see if the software actually works](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). An automated test suite is ideal, but documented manual steps are okay too. But there must be some way for the reviewer to objectively assess whether the software works.

The manual steps in the README sound good then.
