### 2019-11-08 #1828 @stfbnc
@ali-ramadhan Thanks for the comments, I'm now working on resolving the issues and on the documentation.

@mbobra  Can I modify the paper as ali-ramadhan suggested, or should I wait for the second reviewer's comments?
