### 2019-11-08 #1828 @mbobra
@stfbnc Yes, feel free to modify the paper right away! Every time you modify, you comment in this thread with the command `@whedon generate pdf` to ask the JOSS bot Whedon to compile a PDF (so that we can all see the latest draft). This is an iterative process, so there is absolutely no issue with modifying and re-generating the paper as many times as you like.
