### 2019-11-09 stfbnc/fathon#1 @stfbnc
There was an error in the README, try `python  setup.py build_ext --include-dirs=/path/to/GSL/includes --rpath=/path/to/GSL/libs --inplace`
