### 2019-11-09 stfbnc/fathon#1 @stfbnc
In your specific case, `python setup.py build_ext --include-dirs=/usr/include/ --rpath=/usr/lib/ --inplace`
