+++ setup.py
@@ -1,4 +1,3 @@
-from setuptools.command.install import install
 from distutils.core import setup, Extension
 from setuptools import find_packages
 import subprocess
@@ -6,21 +5,15 @@ import os
 from Cython.Build import cythonize
 import numpy
 import platform
+import sys
 from pathlib import Path
 
-home_path = str(Path.home()) ##3.5+ altrimenti from os.path import expanduser - home = expanduser("~")
+home_path =  os.path.expanduser("~")
 
 def gsl_install():
     command = "mkdir -p "+home_path+"/fathonGSL && cd src/gsl_code/ && ./configure --prefix="+home_path+"/fathonGSL && make && make install && cd .. && rm -rf gsl_code"
     process = subprocess.Popen(command, shell=True, cwd=Path(__file__).parent.absolute())
     process.wait()
-
-#class CustomInstall(install):
-#    def run(self):
-#        command = "mkdir -p "+home_path+"/fathonGSL && cd src/gsl_code/ && ./configure --prefix="+home_path+"/fathonGSL && make && make install && cd .. && rm -rf gsl_code"
-#        process = subprocess.Popen(command, shell=True, cwd=Path(__file__).parent.absolute())
-#        process.wait()
-#        install.run(self)
 
 def get_extension(module_name, src_name, current_os):
     sources = [src_name, "src/cLoops.c"]
@@ -47,27 +40,27 @@ def get_extension(module_name, src_name, current_os):
                          extra_link_args=extra_link_args)
 
 if __name__ == '__main__':
-    gsl_install()
+    if sys.version_info[0] == 3:
+        gsl_install()
 
-    running_os = platform.system()
-    if running_os != "Windows":
-        extensions = [get_extension("dfa", "src/dfa.pyx", running_os),
-                      get_extension("dcca", "src/dcca.pyx", running_os),
-                      get_extension("mfdfa", "src/mfdfa.pyx", running_os),
-                      get_extension("ht", "src/ht.pyx", running_os)]
+        running_os = platform.system()
+        if running_os != "Windows":
+            extensions = [get_extension("dfa", "src/dfa.pyx", running_os),
+                          get_extension("dcca", "src/dcca.pyx", running_os),
+                          get_extension("mfdfa", "src/mfdfa.pyx", running_os),
+                          get_extension("ht", "src/ht.pyx", running_os)]
 
-        setup(name="fathon",
-              version="0.1",
-              author="Stefano Bianchi",
-              url="https://github.com/stfbnc/fathon.git",
-              license="GPLv3.0",
-              description="pyhton package for detrended fluctuation analysis (DFA) and related algorithms.",
-              packages=find_packages(),
-              install_requires=["numpy", "cython"],
-              ext_modules=cythonize(extensions, build_dir="build")
-              )
+            setup(name="fathon",
+                  version="0.1",
+                  author="Stefano Bianchi",
+                  url="https://github.com/stfbnc/fathon.git",
+                  license="GPLv3.0",
+                  description="pyhton package for detrended fluctuation analysis (DFA) and related algorithms.",
+                  packages=find_packages(),
+                  install_requires=["numpy", "cython"],
+                  ext_modules=cythonize(extensions, build_dir="build")
+                  )
+        else:
+            print("fathon is not available on Windows yet.")
     else:
-        print("fathon is not available on Windows yet.")
-
-#cmdclass={'install': CustomInstall},
-#include_package_data=True,
+        sys.exit("fathon requires pyhton 3.")
