+++ README.md
@@ -6,34 +6,37 @@ It is mostly written in Cython and C in order to speed up computations.
 
 ## Getting started
 
-To download the code you have two options:
-
-1. From Terminal, run <code>git clone https://github.com/stfbnc/fathon.git</code> 
-2. [Go to the repository](https://github.com/stfbnc/fathon.git) and download the code
-
 ### Prerequisites
+
+At the moment, *fathon* is only available for Linux and macOS.
 
 To make *fathon* work, you need
 
 - Python 3
-- GSL (*GNU Scientific Library*), to perform polynomial fits. You can download GSL [here](https://www.gnu.org/software/gsl/)
+- Cython
 
-Since it is possible to exploit parallel computation, a C compiler with OpenMP support is required under Linux (gcc recommended).
+on both Linux and macOs.
 
-Under macOS, the default C compiler does not support OpenMP. To exploit parallelisation, you have to install (even though not recommended) a C compiler that supports OpenMP, like gcc. See for instance [this](https://stackoverflow.com/questions/54776301/cython-prange-is-repeating-not-parallelizing) and links therein.
+Under Linux only, a C compiler with OpenMP support is required (gcc recommended).
+
+Under macOS, the default C compiler (via Xcode) does not support OpenMP and *fathon* will generally run slower than on Linux. If you want to exploit parallelisation, you have to install another C compiler that supports OpenMP (like gcc). You also have to modifiy the file <code>setup.py</code> present in the *fathon* package:
+
+- see [this](https://stackoverflow.com/questions/54776301/cython-prange-is-repeating-not-parallelizing) and links therein for how to set the C compiler
+- inside <code>setup.py</code>, in function <code>get_extension()</code>, replace the code after the <code>if current_os == "Darwin":</code> with the code after the <code>elif current_os == "Linux":</code>
 
 ### Installing
 
-1. Open Terminal and go to the folder where you have downloaded *fathon*
-2. Enter the *fathon* folder
-3. Run <code>python  setup.py build_ext --include-dirs=/path/to/GSL/includes --rpath=/path/to/GSL/libs --inplace</code>, where <code>/path/to/GSL/includes</code> and <code>/path/to/GSL/libs</code> are the paths where you have installed the GSL includes and libraries respectively
-4. In the *fathon* folder there are now four files corresponding to the compiled Cyhon code
-5. Move the folder *fathon* to the standard location of Python packages (e.g. <code>/your/path/pythonversion/site-packages</code>) or to where you want to use the package
-6. You can now use *fathon* by typing <code>import fathon</code> from Python
+1. Open the Terminal, go to the directory where you want to download the *fathon* package, and run <code>git clone https://github.com/stfbnc/fathon.git</code>
+2. <code>cd fathon</code>
+3. Run <code>python  setup.py build_ext --inplace</code>. This will take a few minutes. It will
+   - install the [GSL (*GNU Scientific Library*)](https://www.gnu.org/software/gsl/) in a folder named <code>fathonGSL</code> in your home directory
+   - compile the Cython scripts directly in the <code>fathon</code> folder
+   - move all the necessary files in the folder <code>fathon</code> created in the last path of your current Python's <code>sys.path</code>
+4. You can now use *fathon* by typing <code>import fathon</code> from Python
 
-## Tests
+## Examples
 
-A jupyter notebook can be used (<code>fathon_example.ipynb</code>) to test the four algorithms provided by the *fathon* package, namely
+A jupyter notebook can be used (<code>fathon_example.ipynb</code>) to play with the four algorithms provided by the *fathon* package, namely
 
[…]
