+++ setup.py
@@ -8,17 +8,23 @@ import platform
 import sys
 from pathlib import Path
 
-home_path =  os.path.expanduser("~")
+#home_path =  os.path.expanduser("~")
+fathon_path = sys.path[-1]+"/fathon"
+gsl_path = fathon_path+"/fathonGSL"
 
 def gsl_install():
-    command = "mkdir -p "+home_path+"/fathonGSL && cd src/gsl_code/ && ./configure --prefix="+home_path+"/fathonGSL && make && make install && cd .. && rm -rf gsl_code"
+    for p in sys.path:
+        process = subprocess.Popen("if [ -d "+p+"/fathon ]; then rm -rf "+p+"/fathon; fi", shell=True, cwd=Path(__file__).parent.absolute())
+        process.wait()
+    #command = "mkdir -p "+home_path+"/fathonGSL && cd src/gsl_code/ && ./configure --prefix="+home_path+"/fathonGSL && make && make install && cd .. && rm -rf gsl_code"
+    command = "mkdir "+fathon_path+" && mkdir "+gsl_path+" && cd src/gsl_code/ && ./configure --prefix="+gsl_path+" && make && make install && cd .. && rm -rf gsl_code"
     process = subprocess.Popen(command, shell=True, cwd=Path(__file__).parent.absolute())
     process.wait()
 
 def get_extension(module_name, src_name, current_os):
     sources = [src_name, "src/cLoops.c"]
-    include_dirs = [numpy.get_include(), home_path+"/fathonGSL/include"]
-    library_dirs = [home_path+"/fathonGSL/lib"]
+    include_dirs = [numpy.get_include(), gsl_path+"/include"]
+    library_dirs = [gsl_path+"/lib"]
     libraries = ["gsl", "gslcblas", "m"]
     extra_compile_args_macos = ["-O2"]
     extra_compile_args_linux = ["-O2", "-fopenmp"]
@@ -40,11 +46,12 @@ def get_extension(module_name, src_name, current_os):
                          extra_link_args=extra_link_args)
 
 def move_fathon():
-    for p in sys.path:
-        process = subprocess.Popen("if [ -d "+p+"/fathon ]; then rm -rf "+p+"/fathon; fi", shell=True, cwd=Path(__file__).parent.absolute())
-        process.wait()
-    fathon_path = sys.path[-1]+"/fathon"
-    mv_fathon = "mkdir "+fathon_path+" && cp __init__.py "+fathon_path+" && cp tsHelper.py "+fathon_path+" && cp README.md "+fathon_path+" && cp LICENSE "+fathon_path+" && mv dcca* "+fathon_path+" && mv dfa* "+fathon_path+" && mv ht* "+fathon_path+" && mv mfdfa* "+fathon_path
+#    for p in sys.path:
+#        process = subprocess.Popen("if [ -d "+p+"/fathon ]; then rm -rf "+p+"/fathon; fi", shell=True, cwd=Path(__file__).parent.absolute())
+#        process.wait()
+    #fathon_path = sys.path[-1]+"/fathon"
+    #mv_fathon = "mkdir "+fathon_path+" && cp __init__.py "+fathon_path+" && cp tsHelper.py "+fathon_path+" && cp README.md "+fathon_path+" && cp LICENSE "+fathon_path+" && mv dcca* "+fathon_path+" && mv dfa* "+fathon_path+" && mv ht* "+fathon_path+" && mv mfdfa* "+fathon_path
[…]
