### 2019-11-10 stfbnc/fathon#2 @stfbnc
@ali-ramadhan I have modified the README.md with the new installation guidelines.
I did not manage to make fathon available via pip yet, but GSL is no more required now and is automatically installed along with fathon in the last of the paths where python searches when a package is imported.

Let me know if you still have problems with the installation.
