### 2019-11-10 stfbnc/fathon#1 @stfbnc
@ali-ramadhan I have now changed the installation guidelines in the README.md, and GSL is no more required.

Let me know if you still have problems with the installation.
