+++ paper/paper.bib
@@ -1,9 +1,19 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+%% Created for Stefano Bianchi at 2019-11-10 20:50:35 +0100 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
 @article{zebende_2011,
 	Author = {Zebende, G.F.},
 	Journal = {Physica A: Statistical Mechanics and its Applications},
 	Number = {4},
 	Pages = {614-618},
-	Title = {DCCA cross-correlation coefficient: quantifying level of  cross-correlation.},
+	Title = {DCCA cross-correlation coefficient: quantifying level of cross-correlation.},
 	Volume = {390},
 	Year = {2011}}
 
@@ -42,24 +52,6 @@
 	Volume = {69},
 	Year = {2004}}
 
-@article{grau_2000,
-	Author = {Grau-Carles, P.},
-	Journal = {Physica A: Statistical Mechanics and its Applications},
-	Number = {3},
-	Pages = {396-404},
-	Title = {Empirical evidence of long-range correlations in  stock returns.},
-	Volume = {287},
-	Year = {2000}}
-
-@article{janosi_1999,
-	Author = {J\`anosi, I.M. and Janecsk\`o, B. and Kondor, I.},
-	Journal = {Physica A: Statistical Mechanics and its Applications},
-	Number = {1},
-	Pages = {111-124},
-	Title = {Statistical analysis of 5s index data of the Budapest Stock Exchange. },
-	Volume = {269},
-	Year = {1999}}
-
 @article{liu_1999,
 	Author = {Liu, Y. and Gopikrishnan, P. and Stanley, H.E.},
 	Journal = {Physical Review E},
@@ -69,24 +61,6 @@
 	Volume = {60},
 	Year = {1999}}
 
-@article{kavasseri_2004,
-	Author = {Kavasseri, R.G. and Nagarajan, R.},
-	Journal = {IEEE Transactions on Circuits and Systems I: Regular Papers},
-	Number = {11},
-	Pages = {2255-2262},
-	Title = {Evidence of crossover phenomena in wind-speed data.},
-	Volume = {51},
-	Year = {2004}}
-
-@article{matsoukas_2000,
-	Author = {Matsoukas, C. and Islam, S. and Rodriguez-Iturbe, I.},
-	Journal = {Journal of Geophysical Research: Atmospheres},
-	Number = {D23},
-	Pages = {29165-29172},
-	Title = {Detrended  fluctuation analysis of rainfall and streamflow time series.},
-	Volume = {105},
-	Year = {2000}}
-
 @article{fraedrich_2003,
 	Author = {Fraedrich, K. and Blender, R.},
 	Journal = {Physical Review Letters},
@@ -95,24 +69,6 @@
 	Title = {Scaling of atmosphere and ocean temperature correlations in observations and climate models.},
 	Volume = {90},
 	Year = {2003}}
-
-@article{talkner_2000,
-	Author = {Talkner, P. and Weber, R.O.},
-	Journal = {Physical Review E},
-	Number = {1},
-	Pages = {150},
-	Title = {Power spectrum and detrended fluctuation analysis: Application to daily temperatures.},
-	Volume = {62},
-	Year = {2000}}
-
-@article{koscielny_1998,
-	Author = {Koscielny-Bunde, E. and Bunde, A. and Havlin, S. and Roman, H.E. and Goldreich, Y. and Schellnhuber, H.J.},
-	Journal = {Physical Review Letters},
-	Number = {3},
-	Pages = {729},
[…]
