+++ paper/paper.md
@@ -12,26 +12,60 @@ authors:
 affiliations:
   - name: Department of Mathematics and Physics, Roma Tre University
     index: 1
-date: 13 October 2019
+date: 10 November 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-Detrended fluctuation analysis (DFA) was first developed by Peng to study memory effects in sequences of DNA [@peng_1993]. Since then, it found various applications in other fields, like geophysics or economy [@varotsos_2007; @koscielny_1998; @talkner_2000; @fraedrich_2003; @matsoukas_2000; @kavasseri_2004; @liu_1999; @janosi_1999; @grau_2000; @ivanov_2004]. DFA is useful in recognising long memory processes by means of the scaling properties of the fluctuation function $F (n) \sim n^H$ . $H$ is the Hurst exponents, and its value can tell if a process (most likely a time series) is persistent or anti-persistent:
+Detrended fluctuation analysis (DFA) was first developed by Peng to study memory effects in sequences of DNA [@peng_1993]. DFA is useful in recognising long memory processes by means of the scaling properties of the fluctuation function $F (n)$. It is calculated starting from the cumulative sum of the process $X(t)$ (most likely a time series), $Y(t) = \sum_{t' = 1}^t [ X(t') - \langle X \rangle ]$, with $\langle X \rangle$ the mean of $X(t)$. $Y(t)$ is then divided into $N_n = \lfloor N / n \rfloor$ non-overlapping time intervals of length $n$. Since the length of the time series is often not a multiple of $n$, some data at the end of $Y(t)$ could be excluded. In order to not neglect them, the same procedure can be repeated starting from the end of $Y(t)$, and the total number of time intervals will amount to $2N_n$. The data in each interval $s$ are fitted with a polynomial $Y_s^{fit}$, and the variance is computed, $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \ [ Y ( ( s - 1 ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = 1, \dots , N_n$, and $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \  [ Y ( N - ( s - N_n ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = N_n + 1, \dots , 2N_n$. Finally, the fluctuation function is defined as
+$$
+F ( n ) = \biggl[ \frac{1}{2 N_n} \sum_{s = 1}^{2 N_n} F^2 ( n,s ) \biggr]^{1 / 2}
+$$
+$F(n)$ is expected to scale as $n^H$ . $H$ is the Hurst exponents, and its value can tell if a process is persistent or anti-persistent:
 
-- $H \in [0.0, 0.5) \rightarrow$ antipersistency;
+- $H \in [0.0, 0.5) \rightarrow$ anti-persistency. The process under study is anti-persistent and tends to decrease (increase) after a previous increase (decrease). An anti-persistent process appears very noisy;
 - $H = 0.5 \rightarrow$ uncorrelated process;
-- $H \in (0.5, 1.0] \rightarrow$ persistency;
-- $H > 1.0$ nonstationary process, stronger long-range correlations are present.
+- $H \in (0.5, 1.0] \rightarrow$ persistency. If a process has been increasing (decreasing) for a period $T$, then it is expected to continue to increase (decrease) for a similar period. Persistent processes show long-range correlations and exhibit relatively little noise;
+- $H > 1.0 \rightarrow$ nonstationary process, stronger long-range correlations are present.
 
[…]
