+++ paper/paper.bib
@@ -30,7 +30,7 @@
 	Author = {Ihlen, E.A.},
 	Journal = {Fractal analysis},
 	Pages = {97},
-	Title = {Introduction to multifractal detrended fluctuation analysis in Matlab.},
+	Title = {Introduction to multifractal detrended fluctuation analysis in {Matlab}.},
 	Volume = {3},
 	Year = {2012}}
 
@@ -75,7 +75,7 @@
 	Journal = {Atmospheric Chemistry and Physics},
 	Number = {3},
 	Pages = {629-634},
-	Title = {Long-term memory effect in the atmospheric CO$_2$ concentration at Mauna Loa.},
+	Title = {Long-term memory effect in the atmospheric {CO}$_2$ concentration at {Mauna} {Loa}.},
 	Volume = {7},
 	Year = {2007}}
 
@@ -84,6 +84,6 @@
 	Journal = {Physical Review E},
 	Number = {5},
 	Pages = {3730},
-	Title = {Finite-size effects on long-range correlations: Implications for analyzing DNA sequences},
+	Title = {Finite-size effects on long-range correlations: Implications for analyzing {DNA} sequences},
 	Volume = {47},
 	Year = {1993}}
