### 2019-11-10 #1828 @stfbnc
@ali-ramadhan I have corrected the paper as you suggested, there is now a theoretical introduction on the algorithms covered by fathon.

For the **State of the field** section:
- the three packages you mentioned only cover the DFA algorithm (and with not all the options provided by fathon), while they do not cover MFDFA, DCCA, and HT.
- the Matlab package is instead more comprehensive, and (besides not being written in Python) does not cover all the features offered by fathon. It does not cover DCCA, is less flexible, and some function's parameters are not considered, like the possibility to calculate the fluctuation function only forward or both forward and backward.
