### 2019-11-11 stfbnc/fathon#2 @ashwinvis
I agree with @ali-ramadhan on the GSL installation approach. Consider improving installation of GSL when available via system manager (for Linux: `apt`, `pacman`, `yum` etc.) and fallback to [conda](https://anaconda.org/conda-forge/gsl) for other platforms / when sudo rights are not available. I would prefer to do

     sudo apt install libgsl-dev

in Ubuntu. The current approach has an added problem that by moving GSL source files, the git repo is dirty after installation.
