### 2019-11-11 stfbnc/fathon#1 @ashwinvis
Hi @stfbnc, I should point out the instructions are not sufficient. Could be improved with the

* As a first step, specify build requirements such as `cython`, `numpy` in `requirements.txt` file, with versions if necessary. Followed by
```sh
pip install -r requirements.txt
```
Better solutions to [specify build requirements as `pyproject.toml` ](https://packaging.python.org/specifications/declaring-build-dependencies/)may be  explored in the future.
* Do not rely on shell commands via `subprocess`. Especially, use of  [`shell=True` is not recommended in the official documentation](https://docs.python.org/3/library/subprocess.html?highlight=subprocess#security-considerations). Atleast use `shutil` and `os.path` / `pathlib`. An even better option be to [use `MANIFEST.IN`](https://packaging.python.org/guides/using-manifest-in/https://packaging.python.org/guides/using-manifest-in/) which would copy LICENSE, README etc. while running `python setup.py install`
