### 2019-11-11 stfbnc/fathon#1 @ashwinvis
I followed your installation instructions in Ubuntu as it is now. But I get:

```sh
❯ python -c 'import fathon'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/avmo/src/sandbox/venv-fathon/lib/python3.8/site-packages/fathon/__init__.py", line 19, in <module>
    from .dfa import DFA
ImportError: libgsl.so.25: cannot open shared object file: No such file or directory
```

I had to run:

```sh
❯ LD_LIBRARY_PATH="../venv-fathon/lib/python3.8/site-packages/fathon/fathonGSL/lib/" python -c 'import fathon'
```

to make it work. Note that I am using a virtual environment.
