+++ setup.py
@@ -6,23 +6,33 @@ from Cython.Build import cythonize
 import numpy
 import platform
 import sys
-from pathlib import Path
+import urllib.request
 
-fathon_path = sys.path[-1]+"/fathon"
-gsl_path = fathon_path+"/fathonGSL"
+main_path = os.path.dirname(os.path.abspath(__file__))
+home_path = os.path.expanduser("~")
 
 def gsl_install():
-    for p in sys.path:
-        process = subprocess.Popen("if [ -d "+p+"/fathon ]; then rm -rf "+p+"/fathon; fi", shell=True, cwd=Path(__file__).parent.absolute())
-        process.wait()
-    command = "mkdir "+fathon_path+" && mkdir "+gsl_path+" && cd src/gsl_code/ && ./configure --prefix="+gsl_path+" && make && make install && cd .. && rm -rf gsl_code"
-    process = subprocess.Popen(command, shell=True, cwd=Path(__file__).parent.absolute())
-    process.wait()
+	process = subprocess.Popen(os.path.join(".", "fathon", "fathon_gsl_install"), cwd=main_path)
+	process.wait()
+	if process.returncode != 0:
+		sys.exit("Failed to install GSL.")
+
+gsl_inc = os.environ.get("GSLINC", None)
+gsl_lib = os.environ.get("GSLLIB", None)
+if gsl_inc is None and gsl_lib is None:
+    urllib.request.urlretrieve("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz", os.path.join(home_path, "gsl-latest.tar.gz"))
+    gsl_install()
+    gsl_inc = "/usr/local/include"
+    gsl_lib = "/usr/local/lib"
+elif gsl_inc is not None and gsl_lib is not None:
+	pass
+else:
+	sys.exit("Both GSLINC and GSLLIB must or must not be given.")
 
 def get_extension(module_name, src_name, current_os):
-    sources = [src_name, "src/cLoops.c"]
-    include_dirs = [numpy.get_include(), gsl_path+"/include"]
-    library_dirs = [gsl_path+"/lib"]
+    sources = [src_name, os.path.join("fathon", "cLoops.c")]
+    include_dirs = [numpy.get_include(), gsl_inc]
+    library_dirs = [gsl_lib]
     libraries = ["gsl", "gslcblas", "m"]
     extra_compile_args_macos = ["-O2"]
     extra_compile_args_linux = ["-O2", "-fopenmp"]
@@ -43,35 +53,28 @@ def get_extension(module_name, src_name, current_os):
                          extra_compile_args=extra_compile_args_linux,
                          extra_link_args=extra_link_args)
 
-def move_fathon():
-    mv_fathon = "cp __init__.py "+fathon_path+" && cp tsHelper.py "+fathon_path+" && cp README.md "+fathon_path+" && cp LICENSE "+fathon_path+" && mv dcca* "+fathon_path+" && mv dfa* "+fathon_path+" && mv ht* "+fathon_path+" && mv mfdfa* "+fathon_path
-    process = subprocess.Popen(mv_fathon, shell=True, cwd=Path(__file__).parent.absolute())
-    process.wait()
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     if sys.version_info[0] == 3:
         running_os = platform.system()
         if running_os != "Windows":
-            gsl_install()
-        
-            extensions = [get_extension("dfa", "src/dfa.pyx", running_os),
-                          get_extension("dcca", "src/dcca.pyx", running_os),
-                          get_extension("mfdfa", "src/mfdfa.pyx", running_os),
-                          get_extension("ht", "src/ht.pyx", running_os)]
[…]
