+++ MANIFEST.in
@@ -1,0 +1,5 @@
+include LICENSE
+recursive-include fathon fathon_gsl_install
+recursive-include fathon *.pyx
+recursive-include fathon *.c
+recursive-include fathon *.h

[…]
