+++ setup.py
@@ -6,7 +6,8 @@ from Cython.Build import cythonize
 import numpy
 import platform
 import sys
-import urllib.request
+import wget
+#import urllib.request
 
 main_path = os.path.dirname(os.path.abspath(__file__))
 home_path = os.path.expanduser("~")
@@ -20,7 +21,8 @@ def gsl_install():
 gsl_inc = os.environ.get("GSLINC", None)
 gsl_lib = os.environ.get("GSLLIB", None)
 if gsl_inc is None and gsl_lib is None:
-    urllib.request.urlretrieve("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz", os.path.join(home_path, "gsl-latest.tar.gz"))
+    #urllib.request.urlretrieve("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz", os.path.join(home_path, "gsl-latest.tar.gz"))
+    wget.download("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz")
     gsl_install()
     gsl_inc = "/usr/local/include"
     gsl_lib = "/usr/local/lib"
@@ -70,7 +72,7 @@ if __name__ == "__main__":
                   license="GPLv3.0",
                   description="pyhton package for detrended fluctuation analysis (DFA) and related algorithms.",
                   packages=find_packages(),
-                  install_requires=["numpy", "cython"],
+                  install_requires=["wget", "numpy", "cython"],
                   ext_modules=cythonize(extensions),
                   package_data={"fathon": ["LICENSE"]},
                   include_package_data=True)
