+++ setup.py
@@ -22,7 +22,7 @@ gsl_inc = os.environ.get("GSLINC", None)
 gsl_lib = os.environ.get("GSLLIB", None)
 if gsl_inc is None and gsl_lib is None:
     #urllib.request.urlretrieve("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz", os.path.join(home_path, "gsl-latest.tar.gz"))
-    wget.download("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz")
+    wget.download("ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz", os.path.join(home_path, "gsl-latest.tar.gz"))
     gsl_install()
     gsl_inc = "/usr/local/include"
     gsl_lib = "/usr/local/lib"
