+++ .travis.yml
@@ -51,8 +51,6 @@ before_install:
   - pip install -U pytest
   - pip install codecov
   - pip3 install -r requirements.txt
-install:
-  - pip install ".[test]"
 script: pytest
 after_success:
   - codecov
