+++ .travis.yml
@@ -21,9 +21,12 @@ matrix:
         - pip3 install -r requirements.txt
         - wget ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz -O $HOME/gsl-latest.tar.gz
         - ./fathon/fathon_gsl_install
+        - pwd
+        - ls -lrt
+      install:
+        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" pip3 install .
       script:
-        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" python3 setup.py build_ext --inplace
-        - python3 -m pytest
+        - pytest
       after_success: python 3 -m codecov
     - name: "Python 3.7.3 on macOS 10.14"
       os: osx
@@ -37,9 +40,12 @@ matrix:
         - pip3 install -r requirements.txt
         - wget ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz -O $HOME/gsl-latest.tar.gz
         - ./fathon/fathon_gsl_install
+        - pwd
+        - ls -lrt
+      install:
+        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" pip3 install .
       script:
-        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" python3 setup.py build_ext --inplace
-        - python3 -m pytest
+        - pytest
       after_success: python 3 -m codecov
     - name: "Python 3.8 on macOS 10.14"
       os: osx
@@ -53,9 +59,12 @@ matrix:
         - pip3 install -r requirements.txt
         - wget ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz -O $HOME/gsl-latest.tar.gz
         - ./fathon/fathon_gsl_install
+        - pwd
+        - ls -lrt
+      install:
+        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" pip3 install .
       script:
-        - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" python3 setup.py build_ext --inplace
-        - python3 -m pytest
+        - pytest
       after_success: python 3 -m codecov
 before_install:
   - python --version
@@ -65,8 +74,11 @@ before_install:
   - pip3 install -r requirements.txt
   - wget ftp://ftp.gnu.org/gnu/gsl/gsl-latest.tar.gz -O $HOME/gsl-latest.tar.gz
   - ./fathon/fathon_gsl_install
+  - pwd
+  - ls -lrt
+install:
+  - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" pip3 install .
 script:
-  - GSLINC="/usr/local/include/" GSLLIB="/usr/local/lib/" python3 setup.py build_ext --inplace
-  - python3 -m pytest
+  - pytest
 after_success:
   - codecov
