### 2019-11-19 875c5a1
minimum numpy version
