+++ README.md
@@ -15,10 +15,16 @@ It is mostly written in Cython and C in order to speed up computations.
 To make *fathon* work, you need:
 
 1. **_Python version:_**
-   - Python 3
+   
+   - Python 3.5 or higher
+   
 2. **_Python packages:_**
-   - Numpy
-   - Cython
+   
+   - wget
+   
+   - numpy (version >= 1.15)
+   - cython
+   
 3. **_OpenMP:_**
    - Linux
      - a C compiler with OpenMP support is required (gcc recommended)
@@ -26,7 +32,9 @@ To make *fathon* work, you need:
      - the default C compiler (via Xcode) does not support OpenMP and *fathon* will generally run slower than on Linux. If you want to exploit parallelisation, you have to install another C compiler that supports OpenMP (like gcc). You also have to modifiy the file <code>setup.py</code> present in the *fathon* package:
        - see [this](https://stackoverflow.com/questions/54776301/cython-prange-is-repeating-not-parallelizing) and links therein for how to set the C compiler
        - inside <code>setup.py</code>, in function <code>get_extension()</code>, replace the code after the <code>if current_os == "Darwin":</code> with the code after the <code>elif current_os == "Linux":</code>
+
 4. **_External libraries:_**
+
    - The [GSL (*GNU Scientific Library*)](https://www.gnu.org/software/gsl/) is required for *fathon* to work. It is not mandatory to install it by yourself since it will be downloaded and installed in your <code>/usr/local/</code> directory during *fathon* installation if you do not have it (see **Installing** section for more informations).
 
 ### Installing
@@ -42,6 +50,13 @@ To install *fathon*, you can choose one of the following (**all paths must be ab
 - If you already have GSL installed, run <code>GSLINC="/path/to/include/" GSLLIB="/path/to/lib/" pip install fathon</code> to install *fathon*. In case you do not know exactly where GSL is installed, you can run <code>find / -name gsl</code> to find the location of <code>/path/to/include/</code> and <code>/path/to/lib/</code>.
 
 :warning: GSL stores the <code>.h</code> files to be included under the <code>.../include/gsl/</code> directory, but the <code>GSLINC</code> path must not go beyond <code>.../include/</code>!
+
+:warning::warning: If you are using a virtual environment you need to add the GSL libraries to your <code>LD_LIBRARY_PATH</code>:
+
+1. Activate the virtual environment;
+2. Run <code>LD_LIBRARY_PATH="/your/GSLLIB/path/:$LD_LIBRARY_PATH"</code>. If you let pip install GSL (the second of the installation options previously listed) <code>/your/GSLLIB/path/</code> is <code>/usr/local/lib/</code>.
+3. Run <code>export LD_LIBRARY_PATH</code>;
+4. Install *fathon*.
 
 ## Examples
 
