+++ requirements.txt
@@ -1,3 +1,3 @@
 wget
-numpy
+numpy>=1.15
 cython
