### 2019-11-20 f1739c1
docs skeleton
