+++ docs/index.rst
@@ -1,0 +1,62 @@
+.. fathon documentation master file, created by
+   sphinx-quickstart on Wed Nov 20 11:37:07 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+fathon (v0.1.2)
+***************
+
+Python package for detrended fluctuation analysis (DFA) and related algorithms
+
+Installation
+============
+
+1. a
+2. b
+3. c
+
+Documentation for the Code
+**************************
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+fathon
+======
+
+.. automodule:: fathon
+   
+tsHelper
+========
+
+description
+
+.. automodule:: fathon.tsHelper
+   :members:
+
+dfa
+===
+.. autoclass:: fathon.dfa
+   :members:
+   
+mfdfa
+=====
+.. autoclass:: fathon.mfdfa
+   :members:
+   
+dcca
+====
+.. autoclass:: fathon.dcca
+   :members:
+   
+ht
+==
+.. autoclass:: fathon.ht
+   :members:
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
