+++ requirements.txt
@@ -1,3 +1,4 @@
 wget
 numpy>=1.15
 cython
+numpydoc
