+++ docs/docstrings_only/fathon/__init__.py
@@ -1,0 +1,25 @@
+#    __init__.py - init for fathon package
+#    Copyright (C) 2019  Stefano Bianchi
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU General Public License as published by
+#    the Free Software Foundation, either version 3 of the License, or
+#    (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU General Public License for more details.
+#
+#    You should have received a copy of the GNU General Public License
+#    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+from .tsHelper import *
+from .dfa import *
+from .mfdfa import *
+from .dcca import *
+from .ht import *
+
+"""A Python package for detrended fluctuation analysis (DFA)
+    and related algorithms.
+"""
