+++ docs/fun_class/fathon.DCCA.rst
@@ -1,0 +1,13 @@
+DCCA
+====
+
+.. currentmodule:: fathon
+
+.. autoclass:: DCCA
+   :show-inheritance:
+
+   .. automethod:: computeFlucVec
+   .. automethod:: computeRho
+   .. automethod:: fitFlucVec
+   .. automethod:: multiFitFlucVec
+   .. automethod:: rhoThresholds
