+++ docs/fun_class/fathon.DFA.rst
@@ -1,0 +1,11 @@
+DFA
+===
+
+.. currentmodule:: fathon
+
+.. autoclass:: DFA
+   :show-inheritance:
+
+   .. automethod:: computeFlucVec
+   .. automethod:: fitFlucVec
+   .. automethod:: multiFitFlucVec

[…]
