### 2019-11-21 stfbnc/fathon#2 @stfbnc
@ali-ramadhan @ashwinvis fathon is now available via `pip` and the user can choose to install GSL by themselves or let `pip` download and install it.
See the README.md for the installation guidelines.
