### 2019-11-21 stfbnc/fathon#1 @stfbnc
@ali-ramadhan @ashwinvis fathon is now available via pip and the user can choose to install GSL by themselves or let pip download and install it.
See the README.md for the installation guidelines.

@ashwinvis I have followed your suggestions. In the readme there are now also instructions for fathon installation in a virtual environment.
