### 2019-11-21 #1828 @stfbnc
@ali-ramadhan @ashwinvis Sorry for the late reply. I have updated the repository. Requirements are now specified, along with the new installation guidelines. I have also added:
 - automated tests on Travis CI
 - readthedocs documentation
 - contributing guidelines
