### 2019-11-21 #1828 @stfbnc
Code coverage is by now 0%, but only because the package is written in cython and not python. I'm still trying to figure out how to make `codecov` consider also `.pyx` or `.so` files.
