### 2019-11-21 stfbnc/fathon#2 @ashwinvis
It would be nice if you can also `git tag ...` the commit corresponding to the PyPI release and push the tag to GitHub.
