### 2019-11-21 #1828 @ashwinvis
@mbobra Thanks for the nudge. I was coordinating with the author to improve the installation and packaging situation https://github.com/stfbnc/fathon/issues/1#issuecomment-552462767 so that we have a version that can be reviewed.

@stfbnc Code coverage is not necessary for the paper, but if you are willing to go that extra mile, it can be fixed by [following this docs](https://cython.readthedocs.io/en/latest/src/tutorial/profiling_tutorial.html#enabling-line-tracing). You need to enable `CYTHON_TRACE=1` and `Cython.Coverage` plugin when you run tests / CI. Do not enable them by default, as they can degrade performance.
