### 2019-12-09 #1828 @stfbnc
@mbobra Actually I already have improved the installation procedure and I have added to the package what the reviewers pointed out was missing.
I was waiting for @ali-ramadhan and @ashwinvis comments on the improved version of the package.
