### 2019-12-17 stfbnc/fathon#2 @ali-ramadhan
@stfbnc Thanks for putting in the work to add fathon to pip!

I'm very sorry for my slow reply.

Currently installing fathon through pip.

It took like ~10 minutes until `setup.py` started running where it seemed like pip was stuck but `htop` showed some intermittent cc/gcc activity so I assumed things were compiling in the background. Was it perhaps compiling some fathon dependencies? If so, might be nice to print a message saying that so users don't think it's stuck or that there's a bug. Most pip packages install in seconds so I was surprised that it took much longer than that.

`setup.py` also took a while to run but there was a message so that was fine.

I thought GSL would be automatically installed but after `pip install fathon` it seemed it's not found:
```
>>> import fathon
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/lib/python3.7/site-packages/fathon/__init__.py", line 19, in <module>
    from .dfa import DFA
ImportError: libgsl.so.25: cannot open shared object file: No such file or directory
```
