### 2019-12-17 stfbnc/fathon#2 @ali-ramadhan
I installed GSL through the Arch Linux package manager and looks like fathon works now!
