### 2019-12-17 stfbnc/fathon#1 @ali-ramadhan
@stfbnc I was able to install fathon on Arch Linux, thank you! A small hiccup with GSL that I mention in #2 but this issue can be closed now.
