### 2019-12-17 #1828 @ali-ramadhan
I'm very sorry for the slow reply on my part.

Thanks @stfbnc for making fathon available through pip. I was able to install it, run the examples, and the tests! Also great work on moving the documentation to Sphinx.

There's just a little hiccup whether the GSL library is actually installed automatically which I mention at https://github.com/stfbnc/fathon/issues/2 but that's a minor issue.

Also thank you @stfbnc for clarifying the difference between fathon and the other packages and revising the software paper. Just gave it another read and it's much better. The paper was helpful in understanding the fathon examples as I'm not super familiar with DFA.

The last paragraph in your software paper reads more like an introduction. I wonder if it would be better to move that paragraph to the very start of the paper (not necessary of course, just a suggestion).

I've updated the review checklist.

My only remaining concern is regarding the tests. Perhaps my ignorance of DFA is showing but I'm actually not sure how fathon is being tested. Take the first test for example:

```python
ts1 = np.loadtxt('./tests/ts1.txt')
ts1 = fathon.toAggregated(ts1)

def test_dfa():
    pydfa = fathon.DFA(ts1)
    n1, F1 = pydfa.computeFlucVec(10, nMax=200, revSeg=True)
    H1, H_int1 = pydfa.fitFlucVec()

    assert math.isclose(H1, 0.7982194289592676)
```

1. What is the data contained in `ts1.txt`? It looks like random numbers but maybe it's real data? Is it e.g. normally distributed?
2. What is being tested? I assume the test is computing a Hurst exponent `H1` and the test is that `H1` comes out correct?
3. Where did the magic number `0.7982194289592676` come from? Is it common to compute Hurst exponents to 16 significant figures? Why do you expect this exact value for `H1`? Does it come from some analytic expression or theory? Or is this value produced by fathon itself, in which case is this a regression test?
