### 2019-12-17 stfbnc/fathon#2 @stfbnc
> It took like ~10 minutes before setup.py started running where it seemed like pip was stuck but htop showed some intermittent cc/gcc activity so I assumed things were compiling in the background. Was it perhaps compiling some fathon dependencies?

It was installing GSL under `/usr/local`.

> I thought GSL would be automatically installed according to the README but after pip install fathon it seemed that it was not installed

Could you please check if GSL includes are under `/usr/local/include/gsl`, and if GSL libraries are under `/usr/local/lib`?

Could you also please check where the package managere installed GSL?
