### 2019-12-17 stfbnc/fathon#2 @ali-ramadhan
Thanks for the quick reply!

Yes you were right, I think pip installed GSL in `/usr/local/` while pacman installed GSL in `/usr/`, not sure if this is an Arch Linux quirk.
