### 2019-12-17 stfbnc/fathon#2 @ashwinvis
> I think pip installed GSL in `/usr/local/`...

This it the right thing to do for installations not done by system package manager. This avoids conflicts. An alternative would be the `/opt` directory.
