### 2019-12-17 stfbnc/fathon#2 @ali-ramadhan
Ah I see, that's good to know. Maybe my local python wasn't looking in `/usr/local/`...

I uninstalled GSL using pacman and then was able to import fathon using the GSL it built with `LD_LIBRARY_PATH=/usr/local/lib`.
```
$ export LD_LIBRARY_PATH=/usr/local/lib
$ python
Python 3.7.4 (default, Oct  4 2019, 06:57:26)
[GCC 9.2.0] on linux
>>> import fathon
>>>
```
