### 2019-12-17 stfbnc/fathon#2 @ashwinvis
Yes, and this step is also pointed out in the readme. I think the issue can be closed for now.

I should note that one would have admin rights only on personal computers, @stfbnc.  One possible improvement for the future is to avoid `sudo ... ` steps and instead:

```bash
./configure --prefix=$HOME/.local
make
make install
```
