### 2019-12-17 stfbnc/fathon#2 @ali-ramadhan
Ah you're right, sorry for not properly reading the note about setting `LD_LIBRARY_PATH`. It even had a big ⚠️ on it!
