### 2019-12-21 stfbnc/fathon#4 @stfbnc
> What is the data contained in `ts1.txt`? It looks like random numbers but maybe it's real data? Is it e.g. normally distributed?

`ts1.txt` in particular contains co2 data after the yearly cycle has been removed. The content  description of the other files has been added in `test_fathon.py`

> What is being tested? I assume the test is computing a Hurst exponent `H1` and the test is that `H1` comes out correct?

Yes, you are right. The description of what is being tested has been added before every test in `test_fathon.py`

> Where did the magic number `0.7982194289592676` come from? Is it common to compute Hurst exponents to 16 significant figures? Why do you expect this exact value for `H1`? Does it come from some analytic expression or theory? Or is this value produced by fathon itself, in which case is this a regression test?

Sorry, it was not specified, but the tests are indeed regression tests. The magic numbers have been produced by fathon and no, it is not common to compute so many significant figures, but I have done that since I do not know the precision of the testing machines on Travis CI.
This information has been added at the beginning of `test_fathon.py`
