+++ paper/paper.md
@@ -18,54 +18,56 @@ bibliography: paper.bib
 
 # Summary
 
-Detrended fluctuation analysis (DFA) was first developed by Peng to study memory effects in sequences of DNA [@peng_1993]. DFA is useful in recognising long memory processes by means of the scaling properties of the fluctuation function $F (n)$. It is calculated starting from the cumulative sum of the process $X(t)$ (most likely a time series), $Y(t) = \sum_{t' = 1}^t [ X(t') - \langle X \rangle ]$, with $\langle X \rangle$ the mean of $X(t)$. $Y(t)$ is then divided into $N_n = \lfloor N / n \rfloor$ non-overlapping time intervals of length $n$. Since the length of the time series is often not a multiple of $n$, some data at the end of $Y(t)$ could be excluded. In order to not neglect them, the same procedure can be repeated starting from the end of $Y(t)$, and the total number of time intervals will amount to $2N_n$. The data in each interval $s$ are fitted with a polynomial $Y_s^{fit}$, and the variance is computed, $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \ [ Y ( ( s - 1 ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = 1, \dots , N_n$, and $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \  [ Y ( N - ( s - N_n ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = N_n + 1, \dots , 2N_n$. Finally, the fluctuation function is defined as
+``fathon`` is a Python package for DFA (Detrended Fluctuation Analysis) and related algorithms.
+
+DFA was first developed by Peng to study memory effects in sequences of DNA [@peng_1993]. DFA is useful in recognising long memory processes by means of the scaling properties of the fluctuation function $F (n)$. $F (n)$ is calculated starting from the cumulative sum of the process $X(t)$ (most likely a time series), $Y(t) = \sum_{t' = 1}^t [ X(t') - \langle X \rangle ]$, with $\langle X \rangle$ the mean of $X(t)$. $Y(t)$ is then divided into $N_n = \lfloor N / n \rfloor$ non-overlapping time intervals of length $n$. Since the length of the time series is often not a multiple of $n$, some data at the end of $Y(t)$ could be excluded. In order to not neglect them, the same procedure can be repeated starting from the end of $Y(t)$, and the total number of time intervals will amount to $2N_n$. The data in each interval $s$ are fitted with a polynomial $Y_s^{fit}$, and the variance is computed, $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \ [ Y ( ( s - 1 ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = 1, \dots , N_n$, and $F^2 ( n,s ) = \frac{1}{n} \sum_{i = 1}^n \  [ Y ( N - ( s - N_n ) n + i ) - Y_s^{fit} (i) ]^2$ for $s = N_n + 1, \dots , 2N_n$. Finally, the fluctuation function is defined as
 $$
 F ( n ) = \biggl[ \frac{1}{2 N_n} \sum_{s = 1}^{2 N_n} F^2 ( n,s ) \biggr]^{1 / 2}
 $$
-$F(n)$ is expected to scale as $n^H$ . $H$ is the Hurst exponents, and its value can tell if a process is persistent or anti-persistent:
+$F(n)$ is expected to scale as $n^H$ . $H$ is the Hurst exponent, and its value can tell if a process is persistent or anti-persistent:
 
[…]
