### 2019-12-21 #1828 @stfbnc
@ali-ramadhan
> The last paragraph in your software paper reads more like an introduction. I wonder if it would be better to move that paragraph to the very start of the paper (not necessary of course, just a suggestion).

I have moved the first sentence at the beginning of the paper, and moved the remainder at the end of the theory. The paragraph contains terms that are described in the theory section, so I think is better to first have the theory.

>  My only remaining concern is regarding the tests.

I have answered your questions in the issue.
