### 2019-12-21 stfbnc/fathon#4 @ali-ramadhan
Thank you for the clarification @stfbnc! Just had a look and `test_fathon.py` is much clearer now so I'll close this issue.
