### 2019-12-21 #1828 @ali-ramadhan
Thanks for the clarification and improvements to the tests @stfbnc!

As all the tests are regression tests however, I am wondering whether that is enough to "cover the core functionality of the software" as per the review criteria.

Regression tests by definition only test that the software's behavior has not changed but by themselves do not test that the software produces the correct answer.

I'd be curious to hear what @mbobra and @ashwinvis's thoughts are on this matter.

I wonder if it would be possible to add tests that cover the core functionality of the software. For example, if you load in a time series of an uncorrelated process (Brownian motion?) do all four methods produce a Hurst exponent of H ≈ 0.5 as expected?
