### 2019-12-22 stfbnc/fathon#4 @ashwinvis
Regression tests are nice to have, but it does not confirm, neither deny that the implementation works. In the PR that I filed just now, I am trying to reproduce a third-party tutorial. Alternatively `fathon` could be tested against some published articles not associated with the developer. Thoughts?
