### 2019-12-22 #1828 @ashwinvis
## Comments

The Python package fathon appears to be a novel and efficient implementation of an analytical tool that was missing from the ecosystem. I congratulate the authorfor the contribution. As of now, the paper gives a good introduction to different algorithms associated with DFA and its variants.

Major revision:

- There is only a single mention describing the "State of the field" with "commonly-used packages". This can be improved with more references, links or description.
- Reproduce a third-party test case to verify that `fathon` implements the  right algorithm. I have tried to attempt this myself in stfbnc/fathon/pull/5.

Minor revision:

- Mention why GSL is a dependency, which I believe is used for fitting polynomials. Maybe also elaborate why this was preferred over `numpy.polyfit`.
- Update link to PyPI release for "Archive" in the manuscript (right now it shows "PENDING")
- Contributing guidelines (`CONTRIBUTING.md`) refers to tags "proposed feature", "new feature" etc. Are these supposed to be plain text / GitHub  labels? If it is the latter, I did not find them.
- The following are optional. It would be beneficial to the reader of the  manuscript to know:
    - The names of the associated module by providing links to the  documentation
    - What missing functionalities are, if any and possible future developmen  goals of this package.

Feel free to ask if any clarifications are needed.
