### 2019-12-27 d155bca
functionality tests
