+++ tests/test_fathon.py
@@ -1,3 +1,141 @@
+import numpy as np
+import scipy.io
+import fathon
+import math
+
+# FUNCTIONALITY TESTS
+# -------------------
+# - Results of the Ihlen's "Introduction to Multifractal Detrended Fluctuation
+#   Analysis in Matlab" are reproduced
+#   (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3366552/)
+# - MATLAB code to reproduce the paper results can be found at:
+#   https://it.mathworks.com/matlabcentral/fileexchange/38262-multifractal-detrended-fluctuation-analyses
+
+#####
+# Time series to test functionality
+# (from Ihlen's MATLAB package "Multifractal detrended fluctuation analyses")
+#####
+mat = scipy.io.loadmat('./tests/mat/fractaldata.mat')
+# white noise
+wn = mat['whitenoise'][:, 0]
+# monofractal time series
+mn = mat['monofractal'][:, 0]
+# multifractal time series
+mf = mat['multifractal'][:, 0]
+#####
+
+# defining parameters and scales as in Ihlen's paper
+nMin = 16
+nMax = 1024
+nScales = 19
+exponents = np.linspace(np.log2(nMin), np.log2(nMax), nScales)
+scales = np.round(np.power(2.0, exponents)).astype(int)
+q_list = [-5, -3, -1, 1, 3, 5]
+
+def get_idxs(vec, scales):
+    idxs = []
+    for s in scales:
+        idxs.append(np.where(vec==s)[0][0])
+    return idxs
+
+#####
+# Functionality test 1
+# It tests DFA for the white noise time series
+# H should be 0.45
+#####
+def test_mat_dfa_wn():
+    w_dfa = fathon.DFA(fathon.toAggregated(wn))
+    n_w, F_w = w_dfa.computeFlucVec(nMin=nMin, nMax=nMax, revSeg=False, nStep=1, polOrd=1)
+    idxs = get_idxs(n_w, scales)
+    n_w = n_w[idxs]
+    F_w = F_w[idxs]
+    H, _ = np.polyfit(np.log2(n_w), np.log2(F_w), 1)
+    assert math.isclose(H, 0.45, rel_tol=1e-2, abs_tol=0)
+
+#####
+# Functionality test 2
+# It tests DFA for the monofractal time series
+# H should be 0.72
+#####
+def test_mat_dfa_mn():
+    mn_dfa = fathon.DFA(fathon.toAggregated(mn))
+    n_mn, F_mn = mn_dfa.computeFlucVec(nMin=nMin, nMax=nMax, revSeg=False, nStep=1, polOrd=1)
+    idxs = get_idxs(n_mn, scales)
+    n_mn = n_mn[idxs]
+    F_mn = F_mn[idxs]
+    H, _ = np.polyfit(np.log2(n_mn), np.log2(F_mn), 1)
+    assert math.isclose(H, 0.72, rel_tol=1e-2, abs_tol=0)
+
+#####
+# Functionality test 3
+# It tests DFA for the multifractal time series
+# H should be 0.77
+#####
+def test_mat_dfa_mf():
+    mf_dfa = fathon.DFA(fathon.toAggregated(mf))
+    n_mf, F_mf = mf_dfa.computeFlucVec(nMin=nMin, nMax=nMax, revSeg=False, nStep=1, polOrd=1)
+    idxs = get_idxs(n_mf, scales)
+    n_mf = n_mf[idxs]
+    F_mf = F_mf[idxs]
+    H, _ = np.polyfit(np.log2(n_mf), np.log2(F_mf), 1)
[…]
