### 2019-12-27 #1828 @stfbnc
@ashwinvis
> There is only a single mention describing the "State of the field" with "commonly-used packages". This can be improved with more references, links or description.

I have added some links and described the differences with fathon

> Reproduce a third-party test case to verify that fathon implements the right algorithm. I have tried to attempt this myself in stfbnc/fathon/pull/5.

I have added tests to reproduce the same results obtained in the Ihlen's paper "Introduction to Multifractal Detrended Fluctuation Analysis in Matlab" (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3366552/)

> Mention why GSL is a dependency, which I believe is used for fitting polynomials. Maybe also elaborate why this was preferred over numpy.polyfit.

I have added a sentence in the README (Prerequisites -> External libraries section)

> Contributing guidelines (CONTRIBUTING.md) refers to tags "proposed feature", "new feature" etc. Are these supposed to be plain text / GitHub labels? If it is the latter, I did not find them.

I have added the missing labels

> Update link to PyPI release for "Archive" in the manuscript (right now it shows "PENDING")

How can I update the link?

> The names of the associated module by providing links to the documentation

Do you mean adding module's name and link in the paper where the specific algorithm is described?
