### 2019-12-27 af41687
functionality tests
