+++ paper/paper.md
@@ -66,8 +66,7 @@ Similar to the standard cross-correlation coefficient, $\rho(n) = 1.0$ indicates
 ``fathon`` implements common operations for all the algorithms described so far, such as the computation of the fluctuation function and its fit in different ranges, the multifractal spectrum, and the detrended cross-correlation index and its confidence intervals. ``fathon`` aims at providing a simple and interactive interface to perform multiple analyses on multiple time series. The Python interface is class-based and user-friendly, while the underlying code is written in Cython and C allowing a fast computation of the algorithms.
 
 ``fathon`` can be useful to all those researchers and data analysts that utilise the methods listed so far, especially in the fields of geophysics, economics, or biophysics. Its purpose is to gather all the algorithms together, in order to have a single package including different methodologies. They are easy to use in their basic or advanced versions, and the user can easily switch between them.
-Undoubtedly, codes for each method have already been individually developed, one above all the Ihlen Matlab package for multifractal detrended fluctuation analysis [@ihlen_2012]. It implements the four algorithms of ``fathon`` even though it does not cover all the methods and functionalities.
-The ``fathon`` package is intended to be more complete, to offer more functions and more flexibility in parameters choice.
+Undoubtedly, codes for each method have already been individually developed, see for instance [mf-dfa](https://github.com/yoshiso/mf-dfa/blob/master/mfdfa.py), [dfa](https://github.com/dokato/dfa), and [detrended_fluctuation](https://raphaelvallat.com/entropy/build/html/generated/entropy.detrended_fluctuation.html). These packages implement only a single algorithm of `fathon` and are less flexible. A most comprehesive package is instead the Ihlen's Matlab package for multifractal detrended fluctuation analysis [@ihlen_2012]. It does not cover DCCA and the computation of the cross-correlation index, is less flexible, and some function's parameters are not considered, like the possibility to calculate the fluctuation function only forward or both forward and backward. The ``fathon`` package is intended to be more complete, gathering different algorithms under a single package. Moreover, `fathon` offers more functions and more flexibility in parameters choice.
 
 # References
 
