### 2019-12-27 #1828 @ashwinvis
> How can I update the link?

I am not sure either. Presumably in the manuscript's metadata that your wrote? The editors should know.

> Do you mean adding module's name and link in the paper where the specific algorithm is described?

Yes. Something like: [`fathon.DFA`](https://fathon.readthedocs.io/en/latest/fun_class/fathon.DFA.html).
