### 2020-01-07 #1828 @mbobra
> > How can I update the link?
>
> I am not sure either. Presumably in the manuscript's metadata that your wrote? The editors should know.

Once this paper passes review, the submitting author will [archive their release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI. The "Archive" link will then lead to this Zenodo entry.
