### 2020-01-07 #1828 @ali-ramadhan
Thanks for adding the six functionality tests @stfbnc. The tests are now automated on Travis CI and the tests cover functionality via functionality and regression tests, which is awesome.

Just ticked off the last item on my checklist so I think fathon should be accepted into JOSS.
