### 2020-01-08 #1828 @stfbnc
@ali-ramadhan @ashwinvis thank you very much for reviewing fathon, and for all the valuable suggestions.

I've recompiled the paper, there were a few typos and now there are also the links to the modules' documentation (as @ashwinvis suggested).
