### 2020-01-08 #1828 @mbobra
@ali-ramadhan @ashwinvis Thank you sincerely for all of the time and effort you dedicated to this review. I truly appreciate that you took it so seriously and gave @stfbnc tons of great feedback that greatly improved the submission. Thank you so much for volunteering to review for JOSS ☀️

@stfbnc Congratulations 🎉 Thank you for working so hard on this submission and integrating all the feedback from the reviewers. I know it was a ton of work and I hope you are happy with the final outcome. We're almost there -- can you please [archive your release in Zenodo](https://guides.github.com/activities/citable-code/), making sure the Zenodo entry has the same title and author list as the paper, and put the resultant Zenodo DOI in your `README.md` file? Thank you!
