### 2020-01-08 #1828 @stfbnc
@mbobra Thank you! I'm really happy with the final outcome!

Here is the Zenodo DOI: https://zenodo.org/badge/latestdoi/214290119
It is also included in the readme.
