### 2020-01-08 #1828 @mbobra
@stfbnc Can you update your `paper.bib` files with the appropriate DOIs, so they appear as links in the paper? For example, the `podobnik_2008` entry in `paper.bib` file should be modified to include the DOI as follows:

```
@article{podobnik_2008,
	Author = {Podobnik, B. and Stanley, H.E.},
	Journal = {Physical Review Letters},
	Number = {8},
	Pages = {084102},
	Title = {Detrended cross-correlation analysis: a new method for analyzing two nonstationary time series.},
	Volume = {100},
	DOI = {10.1016/j.physa.2010.10.022},
	Year = {2008}}
```
