+++ paper/paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% http://bibdesk.sourceforge.net/
 
-%% Created for Stefano Bianchi at 2019-11-10 20:50:35 +0100 
+%% Created for Stefano Bianchi at 2020-01-08 23:31:04 +0100 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -10,6 +10,8 @@
 
 @article{zebende_2011,
 	Author = {Zebende, G.F.},
+	Date-Modified = {2020-01-08 23:28:46 +0100},
+	Doi = {http://dx.doi.org/10.1016/j.physa.2010.10.022},
 	Journal = {Physica A: Statistical Mechanics and its Applications},
 	Number = {4},
 	Pages = {614-618},
@@ -19,6 +21,8 @@
 
 @article{podobnik_2008,
 	Author = {Podobnik, B. and Stanley, H.E.},
+	Date-Modified = {2020-01-08 23:27:23 +0100},
+	Doi = {http://dx.doi.org/10.1103/PhysRevLett.100.084102},
 	Journal = {Physical Review Letters},
 	Number = {8},
 	Pages = {084102},
@@ -28,6 +32,8 @@
 
 @article{ihlen_2012,
 	Author = {Ihlen, E.A.},
+	Date-Modified = {2020-01-08 23:24:23 +0100},
+	Doi = {http://dx.doi.org/10.3389/fphys.2012.00141},
 	Journal = {Fractal analysis},
 	Pages = {97},
 	Title = {Introduction to multifractal detrended fluctuation analysis in {Matlab}.},
@@ -36,6 +42,8 @@
 
 @article{kantelhardt_2002,
 	Author = {Kantelhardt, J.W. and Zschiegner, S.A. and Koscielny-Bunde, E. and Bunde, A. and Havlin, S. and Stanley, H.E.},
+	Date-Modified = {2020-01-08 23:25:32 +0100},
+	Doi = {http://dx.doi.org/10.1016/S0378-4371(02)01383-3},
 	Journal = {Physica A: Statistical Mechanics and its Applications},
 	Number = {1},
 	Pages = {87-114},
@@ -45,6 +53,8 @@
 
 @article{ivanov_2004,
 	Author = {Ivanov, P.C. and Yuen, A. and Podobnik, B. and Lee, Y.},
+	Date-Modified = {2020-01-08 23:30:46 +0100},
+	Doi = {http://dx.doi.org/10.1103/PhysRevE.69.056107},
 	Journal = {Physical Review E},
 	Number = {5},
 	Pages = {056107},
@@ -54,6 +64,8 @@
 
 @article{liu_1999,
 	Author = {Liu, Y. and Gopikrishnan, P. and Stanley, H.E.},
+	Date-Modified = {2020-01-08 23:25:59 +0100},
+	Doi = {http://dx.doi.org/10.1103/PhysRevE.60.1390},
 	Journal = {Physical Review E},
 	Number = {2},
 	Pages = {1390},
@@ -63,6 +75,8 @@
 
 @article{fraedrich_2003,
 	Author = {Fraedrich, K. and Blender, R.},
+	Date-Modified = {2020-01-08 23:23:17 +0100},
+	Doi = {http://dx.doi.org/10.1103/PhysRevLett.90.108501},
 	Journal = {Physical Review Letters},
 	Number = {10},
 	Pages = {108501},
@@ -72,6 +86,8 @@
 
 @article{varotsos_2007,
 	Author = {Varotsos, C. and Assimakopoulos, M.N. and Efstathiou, M.},
[…]
