### 2020-01-08 #1828 @mbobra
@openjournals/joss-eics This paper is accepted and ready for final processing! Nice work, @stfbnc 🎉
