### 2020-01-08 #1828 @danielskatz
Thanks @mbobra - just as a minor correction, let's call it "ready-to-accept" until the final processing is done :)
