### 2020-01-08 #1828 @danielskatz
👋 @stfbnc - please merge https://github.com/stfbnc/fathon/pull/6 or let me know what parts you disagree with
